modify: combine DashboardLayout
